Adding user login

+++ LuxuryFashion-Backend/src/main/java/com/spring/repo/ProductRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Long> {
+public interface ProductRepository extends JpaRepository<Product, Integer> {
 
     List<Product> findByProdStatus(String status);
 //    List<Product> findByProdCategoryAndProdStatus(String prodCategory, String prodStatus);
@@ -36,8 +36,6 @@ public interface ProductRepository extends JpaRepository<Product, Long> {
       )
 """)
     List<Product> searchActiveProducts(@Param("keyword") String keyword);
-
-
 
 }
